Extended API test and added lang check [tracker #12891]
The startup script waited 20 seconds for webconfig, but it's
possible that this might take longer.  The wait time was
extended to 60 seconds.

In addition, the /etc/locale.conf file is loaded before tconsole
is started.  There's a grubenv-to-locale.conf migration that can
occur on first boot, so we might as well use the latest language
configuration found in /etc/locale.conf.